requirements.txt
django
#django-compressor
django-allauth
#mysqlclient
#psycopg2-binary
gunicorn
whitenoise
Pillow==10.0.0
django-htmx==1.16.0
django-tailwind==3.6.0
django-simple-deploy

#django widget tweaks?
psycopg2                      # Added by simple_deploy command.
dj-database-url               # Added by simple_deploy command.